Core - added file path verification for writing XUI and XUR.

+++ XUIHelper.Core/XUR/XUR.cs
@@ -144,8 +144,20 @@ namespace XUIHelper.Core
                     return false;
                 }
 
-                //TODO: Verify file path here!
-                Directory.CreateDirectory(Path.GetDirectoryName(FilePath));
+                if(!XUIHelperCoreUtilities.IsStringValidPath(FilePath))
+                {
+                    Logger?.Here().Error("The provided file path {0} is invalid, returning false.", FilePath);
+                    return false;
+                }
+
+                string? dirName = Path.GetDirectoryName(FilePath);
+                if(string.IsNullOrEmpty(dirName))
+                {
+                    Logger?.Here().Error("The directory name for {0} was null or empty, returning false.", FilePath);
+                    return false;
+                }
+
+                Directory.CreateDirectory(dirName);
                 File.Delete(FilePath);
 
                 int writtenFileSize = 0;
